refactor: move type introspection logic to a new module for better organization and clarity
